Updated the release script to not sign the tag
It appears that creating the release and stripping the release branch
of its new development version is not that great: several times the
release artifacts were created not from the tag but from the dev
version.

The script now shows the commands to issue post release build for
the signing of the release tag.