WICKET-5306 Dynamic extra paramaters not evaluated in ajax submitting behavior in a multipart form
When doing Ajax multipart submit execute the beforeSend handlers and show the indicator just before making the form submit.
This makes it consistent with the non-multipart processing. And additionally there was a chance to show the indicator and then fail to hide it because the form element may not be found